fixed int/Integer problem

+++ src/com/java/phondeux/team/Team.java
@@ -121,7 +121,7 @@ public class Team extends JavaPlugin{
 		
 		eh.RegisterCallback(new EventHandler.EventCallback() {
 			public void run(int parent, int child, String data) {
-				if (th.teamChatter.contains(parent)) th.teamChatter.remove(parent);
+				if (th.teamChatter.contains(parent)) th.teamChatter.remove((Integer) parent);
 			}
 		}, EventHandler.Type.PlayerDeath);
 	}
